Remove duplicate server/src code and merge lobby logic
