--- server/src/game.ts
@@ -1,16 +1,0 @@
-export interface Player {
-  id: string;
-  name: string;
-}
-
-export class Game {
-  private players: Player[] = [];
-
-  addPlayer(player: Player) {
-    this.players.push(player);
-  }
-
-  listPlayers() {
-    return this.players;
-  }
-}
